Added tests that address the issue I'm working on.
The problem I encountered is the following: after user manually resizes
the tree with grow(self, to_size), min and max properties are forgotten.
There is a comment on line 118 in veb/core.py file of grow function
that suggest "there's a bug here with no failing test". I assume that's
*the* bug.